deleted profile settings route, name and picture can be changed on the own profile page
